[Feat] 일기 상세 채소 상태 정렬 변경 (#147)
